requirements.txt
telethon>=1.32.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
backoff>=2.2.1
flask>=2.0.1
gunicorn>=20.1.0 
cryptg 
pyOpenSSL